Create Queue.cpp
Write a program to implement a custom queue using two stacks. The queue should support the following three types of queries:
**Enqueue:** This query type is denoted by "1 x", where x is an element to be enqueued. It means that you need to insert element x at the end of the queue.
**Dequeue:** This query type is denoted by "2". It indicates that you should remove the element at the front of the queue.
**Print Front:** This query type is denoted by "3". It instructs you to print the element at the front of the queue without removing it.
​
**Exercise-1**
input:
1 42,2,1 14,3
​
output:
14
​
**Exercise-2**
Queue
Write a program to implement a custom queue using two stacks. The queue should support the following three types of queries:
Enqueue: This query type is denoted by "1 x", where x is an element to be enqueued. It means that you need to insert element x at the end of the queue.
Dequeue: This query type is denoted by "2". It indicates that you should remove the element at the front of the queue.
Print Front: This query type is denoted by "3". It instructs you to print the element at the front of the queue without removing it.

Exercise-1
input:
1 42,2,1 14,3

output:
14

Exercise-2
input:
1 23,2,1 14,3,2,1 78,3

Output:
14
78
We define a CustomQueue class with two stacks (s1 and s2).
The enqueue method pushes elements onto the first stack.
The dequeue method transfers elements from the first stack to the second stack if the second stack is empty, and then pops from the second stack to simulate dequeuing.
The front method returns the front element by transferring elements from the first stack to the second stack if necessary and then returning the top element of the second stack.
The main function reads the input and performs enqueue, dequeue, and print front operations based on the specified queries.